+++ tests/Integration/Tasks/Set-SamplerTaskVariable.Tests.ps1
@@ -29,6 +29,7 @@ Describe 'Set-SamplerTaskVariable' {
         $originalModuleVersionFolder = $ModuleVersionFolder
         $originalPreReleaseTag = $PreReleaseTag
         $originalBuiltModuleRootScriptPath = $BuiltModuleRootScriptPath
+        $originalBuildInfo = $BuildInfo.Clone()
     }
 
     Context 'When calling the function with parameter AsNewBuild' {
@@ -84,6 +85,82 @@ Describe 'Set-SamplerTaskVariable' {
             $result | Should -Contain ("`tModule Manifest Path (src) = '{0}'" -f (Join-Path -Path $TestDrive -ChildPath 'MyProject/source/MyProject.psd1'))
             $result | Should -Contain "`tModule Version             = '1.0.0-preview'"
             $result | Should -Contain ("`tRelease Notes path         = '{0}'" -f (Join-Path -Path $TestDrive -ChildPath 'MyProject/output/ReleaseNotes.md'))
+        }
+
+        Context 'When BuiltSubModuleDirectory is set on parameter' {
+            BeforeAll {
+                $BuiltModuleSubdirectory = 'SubDir1'
+            }
+
+            AfterAll {
+                $BuiltModuleSubdirectory = $originalBuiltModuleSubdirectory
+            }
+
+            It 'Should return the expected output' {
+                $result = . Sampler\Set-SamplerTaskVariable -AsNewBuild
+
+                Write-Debug ($result | Out-String) -Verbose
+
+                $result | Should -Contain "`tProject Name               = 'MyProject'"
+                $result | Should -Contain ("`tSource Path                = '{0}'" -f (Join-Path -Path $TestDrive -ChildPath 'MyProject/source'))
+                $result | Should -Contain ("`tOutput Directory           = '{0}'" -f (Join-Path -Path $TestDrive -ChildPath 'MyProject/output'))
+                $result | Should -Contain ("`tBuilt Module Subdirectory  = '{0}'" -f (Join-Path -Path $TestDrive -ChildPath 'MyProject/output/SubDir1'))
+                $result | Should -Contain ("`tModule Manifest Path (src) = '{0}'" -f (Join-Path -Path $TestDrive -ChildPath 'MyProject/source/MyProject.psd1'))
+                $result | Should -Contain "`tModule Version             = '1.0.0-preview'"
+                $result | Should -Contain ("`tRelease Notes path         = '{0}'" -f (Join-Path -Path $TestDrive -ChildPath 'MyProject/output/ReleaseNotes.md'))
+            }
+        }
+
+        Context 'When BuiltSubModuleDirectory is set in the build configuration file' {
+            BeforeAll {
+                $BuiltModuleSubdirectory = ''
+                $BuildInfo['BuiltModuleSubdirectory'] = 'SubDir2'
+            }
+
+            AfterAll {
+                $BuiltModuleSubdirectory = $originalBuiltModuleSubdirectory
+                $BuildInfo['BuiltModuleSubdirectory'] = $originalBuildInfo['BuiltModuleSubdirectory']
+            }
+
+            It 'Should return the expected output' {
+                $result = . Sampler\Set-SamplerTaskVariable -AsNewBuild -Verbose
+
+                Write-Debug ($result | Out-String) -Verbose
+
+                $result | Should -Contain "`tProject Name               = 'MyProject'"
+                $result | Should -Contain ("`tSource Path                = '{0}'" -f (Join-Path -Path $TestDrive -ChildPath 'MyProject/source'))
+                $result | Should -Contain ("`tOutput Directory           = '{0}'" -f (Join-Path -Path $TestDrive -ChildPath 'MyProject/output'))
+                $result | Should -Contain ("`tBuilt Module Subdirectory  = '{0}'" -f (Join-Path -Path $TestDrive -ChildPath 'MyProject/output/SubDir2'))
+                $result | Should -Contain ("`tModule Manifest Path (src) = '{0}'" -f (Join-Path -Path $TestDrive -ChildPath 'MyProject/source/MyProject.psd1'))
+                $result | Should -Contain "`tModule Version             = '1.0.0-preview'"
+                $result | Should -Contain ("`tRelease Notes path         = '{0}'" -f (Join-Path -Path $TestDrive -ChildPath 'MyProject/output/ReleaseNotes.md'))
+            }
+        }
+
+        Context 'When BuiltSubModuleDirectory is set in the build configuration file and on parameter' {
+            BeforeAll {
+                $BuiltModuleSubdirectory = 'SubDir3'
+                $BuildInfo['BuiltModuleSubdirectory'] = 'SubDirFile'
+            }
+
+            AfterAll {
+                $BuiltModuleSubdirectory = $originalBuiltModuleSubdirectory
+                $BuildInfo['BuiltModuleSubdirectory'] = $originalBuildInfo['BuiltModuleSubdirectory']
+            }
+
+            It 'Should return the expected output' {
+                $result = . Sampler\Set-SamplerTaskVariable -AsNewBuild
+
+                Write-Debug ($result | Out-String) -Verbose
+
+                $result | Should -Contain "`tProject Name               = 'MyProject'"
+                $result | Should -Contain ("`tSource Path                = '{0}'" -f (Join-Path -Path $TestDrive -ChildPath 'MyProject/source'))
+                $result | Should -Contain ("`tOutput Directory           = '{0}'" -f (Join-Path -Path $TestDrive -ChildPath 'MyProject/output'))
+                $result | Should -Contain ("`tBuilt Module Subdirectory  = '{0}'" -f (Join-Path -Path $TestDrive -ChildPath 'MyProject/output/SubDir3'))
+                $result | Should -Contain ("`tModule Manifest Path (src) = '{0}'" -f (Join-Path -Path $TestDrive -ChildPath 'MyProject/source/MyProject.psd1'))
+                $result | Should -Contain "`tModule Version             = '1.0.0-preview'"
+                $result | Should -Contain ("`tRelease Notes path         = '{0}'" -f (Join-Path -Path $TestDrive -ChildPath 'MyProject/output/ReleaseNotes.md'))
+            }
         }
     }
 
@@ -165,6 +242,7 @@ Describe 'Set-SamplerTaskVariable' {
             $OutputDirectory = $originalOutputDirectory
             $BuiltModuleSubdirectory = $originalBuiltModuleSubdirectory
             $ReleaseNotesPath = $originalReleaseNotesPath
+            $BuildInfo = $originalBuildInfo.Clone()
         }
 
         It 'Should return the expected output' {
@@ -192,6 +270,127 @@ Describe 'Set-SamplerTaskVariable' {
             $result | Should -Contain "`tPre-release Tag            = 'preview'"
             $result | Should -Contain ("`tBuilt Module Root Script   = '{0}'" -f (Join-Path -Path $TestDrive -ChildPath 'MyProject\output\MyProject\1.0.0-preview\MyProject.psm1'))
         }
+
+        Context 'When BuiltSubModuleDirectory is set on parameter' {
+            BeforeAll {
+                $BuiltModuleSubdirectory = 'SubDir4'
+                $mockBuiltModuleManifest = Join-Path -Path $TestDrive -ChildPath 'MyProject/output/SubDir4/MyProject/1.0.0-preview/MyProject.psd1'
+                $mockBuiltModuleBase = Join-Path -Path $TestDrive -ChildPath 'MyProject/output/SubDir4/MyProject/1.0.0-preview'
+                $mockBuiltModuleRootScriptPath = Join-Path -Path $TestDrive -ChildPath 'MyProject/output/SubDir4/MyProject/1.0.0-preview/MyProject.psm1'
+            }
+
+            AfterAll {
+                $BuiltModuleSubdirectory = $originalBuiltModuleSubdirectory
+            }
+
+            It 'Should return the expected output' {
+                <#
+                    Since Sampler adds its own alias in build.ps1 that does not point
+                    to the built module's Set-SamplerTaskVariable we must point
+                    out that the alias to test is the one in the module.
+                #>
+                $result = . Sampler\Set-SamplerTaskVariable
+
+                Write-Debug ($result | Out-String) -Verbose
+
+                $result | Should -Contain "`tProject Name               = 'MyProject'"
+                $result | Should -Contain ("`tSource Path                = '{0}'" -f (Join-Path -Path $TestDrive -ChildPath 'MyProject/source'))
+                $result | Should -Contain ("`tOutput Directory           = '{0}'" -f (Join-Path -Path $TestDrive -ChildPath 'MyProject/output'))
+                $result | Should -Contain ("`tBuilt Module Subdirectory  = '{0}'" -f (Join-Path -Path $TestDrive -ChildPath 'MyProject/output/SubDir4'))
+                $result | Should -Contain ("`tModule Manifest Path (src) = '{0}'" -f (Join-Path -Path $TestDrive -ChildPath 'MyProject/source/MyProject.psd1'))
+                $result | Should -Contain "`tModule Version             = '1.0.0-preview'"
+                $result | Should -Contain ("`tRelease Notes path         = '{0}'" -f (Join-Path -Path $TestDrive -ChildPath 'MyProject/output/ReleaseNotes.md'))
+
+                $result | Should -Contain "`tVersioned Output Directory = 'True'"
+                $result | Should -Contain ("`tBuilt Module Manifest      = '{0}'" -f (Join-Path -Path $TestDrive -ChildPath 'MyProject\output\SubDir4\MyProject\1.0.0-preview\MyProject.psd1'))
+                $result | Should -Contain ("`tBuilt Module Base          = '{0}'" -f (Join-Path -Path $TestDrive -ChildPath 'MyProject\output\SubDir4\MyProject\1.0.0-preview'))
+                $result | Should -Contain "`tModule Version Folder      = '1.0.0'"
+                $result | Should -Contain "`tPre-release Tag            = 'preview'"
+                $result | Should -Contain ("`tBuilt Module Root Script   = '{0}'" -f (Join-Path -Path $TestDrive -ChildPath 'MyProject\output\SubDir4\MyProject\1.0.0-preview\MyProject.psm1'))
+            }
+        }
+
+        Context 'When BuiltSubModuleDirectory is set in the build configuration file' {
+            BeforeAll {
+                $BuiltModuleSubdirectory = ''
+                $BuildInfo['BuiltModuleSubdirectory'] = 'SubDir5'
+                $mockBuiltModuleManifest = Join-Path -Path $TestDrive -ChildPath 'MyProject/output/SubDir5/MyProject/1.0.0-preview/MyProject.psd1'
+                $mockBuiltModuleBase = Join-Path -Path $TestDrive -ChildPath 'MyProject/output/SubDir5/MyProject/1.0.0-preview'
+                $mockBuiltModuleRootScriptPath = Join-Path -Path $TestDrive -ChildPath 'MyProject/output/SubDir5/MyProject/1.0.0-preview/MyProject.psm1'
+            }
+
+            AfterAll {
+                $BuiltModuleSubdirectory = $originalBuiltModuleSubdirectory
+                $BuildInfo['BuiltModuleSubdirectory'] = $originalBuildInfo['BuiltModuleSubdirectory']
+            }
+
+            It 'Should return the expected output' {
+                <#
+                    Since Sampler adds its own alias in build.ps1 that does not point
+                    to the built module's Set-SamplerTaskVariable we must point
+                    out that the alias to test is the one in the module.
+                #>
+                $result = . Sampler\Set-SamplerTaskVariable -Verbose
+
+                Write-Debug ($result | Out-String) -Verbose
+
+                $result | Should -Contain "`tProject Name               = 'MyProject'"
+                $result | Should -Contain ("`tSource Path                = '{0}'" -f (Join-Path -Path $TestDrive -ChildPath 'MyProject/source'))
+                $result | Should -Contain ("`tOutput Directory           = '{0}'" -f (Join-Path -Path $TestDrive -ChildPath 'MyProject/output'))
+                $result | Should -Contain ("`tBuilt Module Subdirectory  = '{0}'" -f (Join-Path -Path $TestDrive -ChildPath 'MyProject/output/SubDir5'))
+                $result | Should -Contain ("`tModule Manifest Path (src) = '{0}'" -f (Join-Path -Path $TestDrive -ChildPath 'MyProject/source/MyProject.psd1'))
+                $result | Should -Contain "`tModule Version             = '1.0.0-preview'"
+                $result | Should -Contain ("`tRelease Notes path         = '{0}'" -f (Join-Path -Path $TestDrive -ChildPath 'MyProject/output/ReleaseNotes.md'))
+
+                $result | Should -Contain "`tVersioned Output Directory = 'True'"
+                $result | Should -Contain ("`tBuilt Module Manifest      = '{0}'" -f (Join-Path -Path $TestDrive -ChildPath 'MyProject\output\SubDir5\MyProject\1.0.0-preview\MyProject.psd1'))
+                $result | Should -Contain ("`tBuilt Module Base          = '{0}'" -f (Join-Path -Path $TestDrive -ChildPath 'MyProject\output\SubDir5\MyProject\1.0.0-preview'))
+                $result | Should -Contain "`tModule Version Folder      = '1.0.0'"
+                $result | Should -Contain "`tPre-release Tag            = 'preview'"
+                $result | Should -Contain ("`tBuilt Module Root Script   = '{0}'" -f (Join-Path -Path $TestDrive -ChildPath 'MyProject\output\SubDir5\MyProject\1.0.0-preview\MyProject.psm1'))
+            }
+        }
+
+        Context 'When BuiltSubModuleDirectory is set in the build configuration file and on parameter' {
+            BeforeAll {
+                $BuiltModuleSubdirectory = 'SubDir6'
+                $BuildInfo['BuiltModuleSubdirectory'] = 'SubDirFile'
+                $mockBuiltModuleManifest = Join-Path -Path $TestDrive -ChildPath 'MyProject/output/SubDir6/MyProject/1.0.0-preview/MyProject.psd1'
+                $mockBuiltModuleBase = Join-Path -Path $TestDrive -ChildPath 'MyProject/output/SubDir6/MyProject/1.0.0-preview'
+                $mockBuiltModuleRootScriptPath = Join-Path -Path $TestDrive -ChildPath 'MyProject/output/SubDir6/MyProject/1.0.0-preview/MyProject.psm1'
+            }
+
+            AfterAll {
+                $BuiltModuleSubdirectory = $originalBuiltModuleSubdirectory
+                $BuildInfo['BuiltModuleSubdirectory'] = $originalBuildInfo['BuiltModuleSubdirectory']
+            }
+
+            It 'Should return the expected output with parameter priority' {
+                <#
+                    Since Sampler adds its own alias in build.ps1 that does not point
+                    to the built module's Set-SamplerTaskVariable we must point
+                    out that the alias to test is the one in the module.
+                #>
+                $result = . Sampler\Set-SamplerTaskVariable
+
+                Write-Debug ($result | Out-String) -Verbose
+
+                $result | Should -Contain "`tProject Name               = 'MyProject'"
+                $result | Should -Contain ("`tSource Path                = '{0}'" -f (Join-Path -Path $TestDrive -ChildPath 'MyProject/source'))
+                $result | Should -Contain ("`tOutput Directory           = '{0}'" -f (Join-Path -Path $TestDrive -ChildPath 'MyProject/output'))
+                $result | Should -Contain ("`tBuilt Module Subdirectory  = '{0}'" -f (Join-Path -Path $TestDrive -ChildPath 'MyProject/output/SubDir6'))
+                $result | Should -Contain ("`tModule Manifest Path (src) = '{0}'" -f (Join-Path -Path $TestDrive -ChildPath 'MyProject/source/MyProject.psd1'))
+                $result | Should -Contain "`tModule Version             = '1.0.0-preview'"
+                $result | Should -Contain ("`tRelease Notes path         = '{0}'" -f (Join-Path -Path $TestDrive -ChildPath 'MyProject/output/ReleaseNotes.md'))
+
+                $result | Should -Contain "`tVersioned Output Directory = 'True'"
+                $result | Should -Contain ("`tBuilt Module Manifest      = '{0}'" -f (Join-Path -Path $TestDrive -ChildPath 'MyProject\output\SubDir6\MyProject\1.0.0-preview\MyProject.psd1'))
+                $result | Should -Contain ("`tBuilt Module Base          = '{0}'" -f (Join-Path -Path $TestDrive -ChildPath 'MyProject\output\SubDir6\MyProject\1.0.0-preview'))
+                $result | Should -Contain "`tModule Version Folder      = '1.0.0'"
+                $result | Should -Contain "`tPre-release Tag            = 'preview'"
+                $result | Should -Contain ("`tBuilt Module Root Script   = '{0}'" -f (Join-Path -Path $TestDrive -ChildPath 'MyProject\output\SubDir6\MyProject\1.0.0-preview\MyProject.psm1'))
+            }
+        }
     }
 
     It 'Should have set all values back to original pipeline values' {
@@ -209,5 +408,9 @@ Describe 'Set-SamplerTaskVariable' {
         $ModuleVersionFolder | Should -Be $originalModuleVersionFolder
         $PreReleaseTag | Should -Be $originalPreReleaseTag
         $BuiltModuleRootScriptPath | Should -Be $originalBuiltModuleRootScriptPath
+        foreach ($buildKey in $BuildInfo.Keys)
+        {
+            $BuildInfo[$buildKey] | Should -Be $originalBuildInfo[$buildKey]
+        }
     }
 }