+++ CHANGELOG.md
@@ -11,7 +11,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Added support of `BuiltSubDirectoryDirectory` in build configuration files ([issue #299](https://github.com/gaelcolas/Sampler/issues/299)) .
+- Added support of `BuiltSubDirectoryDirectory` in build configuration files
+ ([issue #299](https://github.com/gaelcolas/Sampler/issues/299)).
+
+### Fixed
+
+- Removed `$BuiltModuleSubdirectory` definition in the `begin` bloc of `build.ps1`
+ template ([issue #299](https://github.com/gaelcolas/Sampler/issues/299)).
 
 ### Fixed